Ponto DOUBLE, 1ptfuga meeeeeeé

+++ include/Ponto.h
@@ -5,12 +5,12 @@
 
 class Ponto {
 public:
-    int x, y, z, m;
-    Ponto(int x, int y);
-    Ponto(int x, int y, int z);
-    Ponto(int x, int y, int z, int m);
+    double x, y, z, m;
+    Ponto(double x, double y);
+    Ponto(double x, double y, double z);
+    Ponto(double x, double y, double z, double m);
 
-    int& operator[] (int idx);
+    double& operator[] (int idx);
     friend std::ostream& operator<< (std::ostream& strm, const Ponto& p);
 
 private:
